Add functional AFN MCTS testing
This commit adds testing to ensure that we are actually
approximating the flow rather than some other value. I have
successfully demonstrated that the flow is correctly approximated
given enough simulations and the ground truth QF values as prior
logits. Note that we don't have access to the ground truth QF
values during training so we will have to rely on changing these
as the search tree is expanded and more information is gleaned. I
am also seeing some issues with the degree of exploration that
this search tree seems to be able to achieve. It looks as though
the tree heavily exploits a few actions. This might not be useful
during a larger training process since early in training we'd like
to explore more I would think.

Danilo