Add new tests and dynamic QF updates
This commit adds new tests, whereby we vary both the noise added to
each flow estimate and the number of simulations. These tests are
still limited by the random number generator, which seems to be
slightly bugged since each batch index seems to produce the exact
same flow estimates. I have not yet investigated this bug.

The dynamic QF update basically just says that we are updating the
flows constantly in the backward() function. Importantly, it seems
that we get much better, and faster-converging flow estimates this
way.

Danilo