Add working constant estimation.
This commit fixes an NaN bug that was caused by doing the
following arithmetic operation: 0 / 0. The issue: child visitation
statistics were not being updated on leaf or terminal nodes,
causing NaNs to be backpropagated through the tree. Simple fix was
to use the Jax equivalent of an if-statement. Testing on our
simple binary tree revealed that it should work perfectly fine,
but may not reduce errors to the extent that was hoped. It remains
to be seen how well it works with predictor networks. This testing
is yet to be done.

Danilo