Comment stuff out
This commit serves as final notice that I will not be using the
corrections branch anymore. The reason: I believe that doing the
flow computations from the child's point of view leads to
numerical instabilities, which causes the large spike in errors that
I was observing in the error and KL graphs. These pieces of information
combined have made me confident in the other method.

Danilo