Add bug fix for afn_mctx: post-terminal states
This bug fix addresses an issue that arises when using MCTS with PGX. In
PGX, you will loop over the same environment even after the initial run
has terminated, using an auto reset feature to get back to the initial
state. The problem is that the search tree is not designed to allow for
restarts, nor does it aid the search in any way: if you loop back to the
start of an environment, then you lose all the state information you had
and have no access to the states around the starting state of the
search. Now, the post-terminal performance of MCTS will allow for auto
reset and for post-terminal value cancellation.

Hopefully this fixes the issues.

Danilo