Storing legal action masking
This commit is to store some progress in legal action masking that
I have since realised that I didn't need to make. If in the future
it is needed, I will keep it here. The only change I really needed
was to change the initialisation value of children_values...

Danilo