Bug fix for log-sum-exp
This commit fixes a bug with log-sum-exp on the initialisation values
for children_values. When these were initialised originally, they took
values of zero, which, in an LSE, become a sum over ones. Which is bad
if you're trying to predict something with the other non-zero values...

I have also removed updates for the child_prior_logits since this is not
something that I should / need to update.

Danilo