Bug fix on post-terminal node values
This commit fixes a hidden issue with post-terminal search states. The
basic problem was that node values of 0 were being propagated up to
parents that should not have any children, causing their value
predictions to be slightly off. This would then cascade into a harder
problem in larger trees.

The fix was to set children_rewards to the min value and explicitly
check for cases where infinite values would be produced.

Danilo